\name{crime_visulization-package}
\alias{crime_visulization-package}
\alias{crime_visulization}
\docType{package}
\title{
\packageTitle{crime_visulization}
}
\description{
\packageDescription{crime_visulization}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{crime_visulization}
\packageIndices{crime_visulization}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{crime_visulization}

Maintainer: \packageMaintainer{crime_visulization}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
